requirements.txt
Pillow
requests
numpy